refactor: refactor ExpectedFelt from virtualMachine to primitives
