+++ src/builtins/rangeCheck.ts
@@ -1,7 +1,7 @@
 import { ExpectedOffset, RangeCheckOutOfBounds } from 'errors/builtins';
 import { BuiltinHandler } from './builtin';
 import { isFelt } from 'primitives/segmentValue';
-import { ExpectedFelt } from 'errors/virtualMachine';
+import { ExpectedFelt } from 'errors/primitives';
 
 export const rangeCheckHandler = (boundExponent: bigint): BuiltinHandler => {
   return {
@@ -9,7 +9,7 @@ export const rangeCheckHandler = (boundExponent: bigint): BuiltinHandler => {
       if (isNaN(Number(prop))) throw new ExpectedOffset();
 
       const offset = Number(prop);
-      if (!isFelt(newValue)) throw new ExpectedFelt();
+      if (!isFelt(newValue)) throw new ExpectedFelt(newValue);
       if (newValue.toBigInt() >> boundExponent !== 0n)
         throw new RangeCheckOutOfBounds();
